+++ BMKit/Sources/BMKit/Session.swift
@@ -1,5 +1,5 @@
 //
-//  Credentials.swift
+//  Session.swift
 //  BMKit
 //
 //  Created by Vincent Tourraine on 13/03/2020.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-public struct Credentials: Codable {
+public struct Session: Codable {
     public let token: String
     public let settingsToken: String
     public let userIdentifier: String