+++ bm/Login/LoginViewController.swift
@@ -132,9 +132,9 @@ class LoginViewController: UIViewController, UITextFieldDelegate {
 
         URLSession.shared.connect(username: subscriberNumber, password: password) { result in
             switch result {
-            case .success(let credentials):
-                credentials.save(to: Credentials.defaultKeychain())
-                self.fetchItems(with: credentials)
+            case .success(let session):
+                session.save(to: Session.defaultKeychain())
+                self.fetchItems(with: session)
 
             case .failure(let error):
                 self.configure(loading: false)
@@ -152,9 +152,9 @@ class LoginViewController: UIViewController, UITextFieldDelegate {
         dismiss(animated: true, completion: nil)
     }
 
-    private func fetchItems(with credentials: Credentials) {
-        let session = URLSession.shared
-        session.fetchItems(with: credentials) { result in
+    private func fetchItems(with session: Session) {
+        let urlSession = URLSession.shared
+        urlSession.fetchItems(with: session) { result in
             self.configure(loading: false)
 
             switch result {